Add BaseWriterAgent class and its methods
