+++ src/agents/coder/coder.py
@@ -2,39 +2,22 @@ import os
 import time
 from typing import Dict, List, Union
 
-from jinja2 import BaseLoader, Environment
-
-from src.config import Config
-from src.llm import LLM
+from src.agents import BaseWriterAgent
 from src.state import AgentState
 
-PROMPT = (
-    open(
-        os.path.join(os.path.dirname(__file__), "prompt.jinja2"), "r", encoding="utf-8"
-    )
-    .read()
-    .strip()
-)
 
+class Coder(BaseWriterAgent):
+    """Coder agent class"""
 
-
-class Coder:
-    def __init__(self, base_model: str):
-        config = Config()
-        self.project_dir = config.get_projects_dir()
-
-        self.llm = LLM(model_id=base_model)
-
-    def render(
-        self, step_by_step_plan: str, user_context: str, search_results: dict
-    ) -> str:
-        env = Environment(loader=BaseLoader())
-        template = env.from_string(PROMPT)
-        return template.render(
-            step_by_step_plan=step_by_step_plan,
-            user_context=user_context,
-            search_results=search_results,
+    _prompt = (
+        open(
+            os.path.join(os.path.dirname(__file__), "prompt.jinja2"),
+            "r",
+            encoding="utf-8",
         )
+        .read()
+        .strip()
+    )
 
     def validate_response(self, response: str) -> Union[List[Dict[str, str]], bool]:
         response = response.strip()
@@ -67,31 +50,8 @@ class Coder:
 
         return result
 
-    def save_code_to_project(self, response: List[Dict[str, str]], project_name: str):
-        file_path_dir = None
-        project_name = project_name.lower().replace(" ", "-")
-
-        for file in response:
-            file_path = f"{self.project_dir}/{project_name}/{file['file']}"
-            file_path_dir = file_path[: file_path.rfind("/")]
-            os.makedirs(file_path_dir, exist_ok=True)
-
-            with open(file_path, "w") as f:
-                f.write(file["code"])
-
-        return file_path_dir
-
-    def get_project_path(self, project_name: str):
-        project_name = project_name.lower().replace(" ", "-")
-        return f"{self.project_dir}/{project_name}"
-
-    def response_to_markdown_prompt(self, response: List[Dict[str, str]]) -> str:
-        response = "\n".join(
-            [f"File: `{file['file']}`:\n```\n{file['code']}\n```" for file in response]
-        )
-        return f"~~~\n{response}\n~~~"
-
     def emulate_code_writing(self, code_set: list, project_name: str):
+        """Emulate the code writing process"""
         for current_file in code_set:
             file = current_file["file"]
             code = current_file["code"]
@@ -115,18 +75,20 @@ class Coder:
         search_results: dict,
         project_name: str,
     ) -> str:
-        prompt = self.render(step_by_step_plan, user_context, search_results)
+        prompt = self.render(
+            step_by_step_plan=step_by_step_plan,
+            user_context=user_context,
+            search_results=search_results,
+        )
+
         response = self.llm.inference(prompt, project_name)
 
         valid_response = self.validate_response(response)
 
         while not valid_response:
-            print("Invalid response from the model, trying again...")
             return self.execute(
                 step_by_step_plan, user_context, search_results, project_name
             )
-
-        print(valid_response)
 
         self.emulate_code_writing(valid_response, project_name)
