+++ src/agents/patcher/patcher.py
@@ -2,45 +2,22 @@ import os
 import time
 from typing import Dict, List, Union
 
-from jinja2 import BaseLoader, Environment
-
-from src.config import Config
-from src.llm import LLM
 from src.state import AgentState
-
-PROMPT = (
-    open(
-        os.path.join(os.path.dirname(__file__), "prompt.jinja2"), "r", encoding="utf-8"
-    )
-    .read()
-    .strip()
-)
+from src.agents import BaseWriterAgent
 
 
-class Patcher:
-    def __init__(self, base_model: str):
-        config = Config()
-        self.project_dir = config.get_projects_dir()
+class Patcher(BaseWriterAgent):
+    """Patcher agent class"""
 
-        self.llm = LLM(model_id=base_model)
-
-    def render(
-        self,
-        conversation: list,
-        code_markdown: str,
-        commands: list,
-        error: str,
-        system_os: str,
-    ) -> str:
-        env = Environment(loader=BaseLoader())
-        template = env.from_string(PROMPT)
-        return template.render(
-            conversation=conversation,
-            code_markdown=code_markdown,
-            commands=commands,
-            error=error,
-            system_os=system_os,
+    _prompt = (
+        open(
+            os.path.join(os.path.dirname(__file__), "prompt.jinja2"),
+            "r",
+            encoding="utf-8",
         )
+        .read()
+        .strip()
+    )
 
     def validate_response(self, response: str) -> Union[List[Dict[str, str]], bool]:
         response = response.strip()
@@ -73,30 +50,6 @@ class Patcher:
 
         return result
 
-    def save_code_to_project(self, response: List[Dict[str, str]], project_name: str):
-        file_path_dir = None
-        project_name = project_name.lower().replace(" ", "-")
-
-        for file in response:
-            file_path = f"{self.project_dir}/{project_name}/{file['file']}"
-            file_path_dir = file_path[: file_path.rfind("/")]
-            os.makedirs(file_path_dir, exist_ok=True)
-
-            with open(file_path, "w") as f:
-                f.write(file["code"])
-
-        return file_path_dir
-
-    def get_project_path(self, project_name: str):
-        project_name = project_name.lower().replace(" ", "-")
-        return f"{self.project_dir}/{project_name}"
-
-    def response_to_markdown_prompt(self, response: List[Dict[str, str]]) -> str:
-        response = "\n".join(
-            [f"File: `{file['file']}`:\n```\n{file['code']}\n```" for file in response]
-        )
-        return f"~~~\n{response}\n~~~"
-
     def emulate_code_writing(self, code_set: list, project_name: str):
         for file in code_set:
             file = file["file"]
@@ -119,7 +72,13 @@ class Patcher:
         system_os: dict,
         project_name: str,
     ) -> str:
-        prompt = self.render(conversation, code_markdown, commands, error, system_os)
+        prompt = self.render(
+            conversation=conversation,
+            code=code_markdown,
+            commands=commands,
+            error=error,
+            system_os=system_os,
+        )
         response = self.llm.inference(prompt, project_name)
 
         valid_response = self.validate_response(response)
